ccip-precommit and fix token approval bug where tokenAmounts includes the non-native fee token

+++ contracts/src/v0.8/ccip/test/production-examples/CCIPSenderTest.t.sol
@@ -7,109 +7,108 @@ import {ICCIPClientBase} from "../../production-examples/interfaces/ICCIPClientB
 import {Client} from "../../libraries/Client.sol";
 import {EVM2EVMOnRampSetup} from "../onRamp/EVM2EVMOnRampSetup.t.sol";
 
-import {IRouterClient} from "../../interfaces/IRouterClient.sol";
 import {IERC20} from "../../../vendor/openzeppelin-solidity/v4.8.3/contracts/token/ERC20/IERC20.sol";
+import {IRouterClient} from "../../interfaces/IRouterClient.sol";
 
-contract CCIPSenderTest is EVM2EVMOnRampSetup { 
-    event MessageFailed(bytes32 indexed messageId, bytes reason);
-    event MessageSucceeded(bytes32 indexed messageId);
-    event MessageRecovered(bytes32 indexed messageId);
+contract CCIPSenderTest is EVM2EVMOnRampSetup {
+  event MessageFailed(bytes32 indexed messageId, bytes reason);
+  event MessageSucceeded(bytes32 indexed messageId);
+  event MessageRecovered(bytes32 indexed messageId);
 
-    CCIPSender internal s_sender;
+  CCIPSender internal s_sender;
 
-    function setUp() public virtual override {
-        EVM2EVMOnRampSetup.setUp();
+  function setUp() public virtual override {
+    EVM2EVMOnRampSetup.setUp();
 
-        s_sender = new CCIPSender(address(s_sourceRouter));
-        s_sender.enableChain(DEST_CHAIN_SELECTOR, abi.encode(address(s_sender)), "");
-    }
+    s_sender = new CCIPSender(address(s_sourceRouter));
+    s_sender.enableChain(DEST_CHAIN_SELECTOR, abi.encode(address(s_sender)), "");
+  }
 
-    function test_ccipSend_withNonNativeFeetoken_andDestTokens() public {
-        address token = address(s_sourceFeeToken);
-        uint256 amount = 111333333777;
-        Client.EVMTokenAmount[] memory destTokenAmounts = new Client.EVMTokenAmount[](1);
-        destTokenAmounts[0] = Client.EVMTokenAmount({token: token, amount: amount});
+  function test_ccipSend_withNonNativeFeetoken_andDestTokens() public {
+    address token = address(s_sourceFeeToken);
+    uint256 amount = 111333333777;
+    Client.EVMTokenAmount[] memory destTokenAmounts = new Client.EVMTokenAmount[](1);
+    destTokenAmounts[0] = Client.EVMTokenAmount({token: token, amount: amount});
 
-        // Make sure we give the receiver contract enough tokens like CCIP would.
-        IERC20(token).approve(address(s_sender), type(uint).max);
+    // Make sure we give the receiver contract enough tokens like CCIP would.
+    IERC20(token).approve(address(s_sender), type(uint256).max);
 
-        Client.EVM2AnyMessage memory message = Client.EVM2AnyMessage({
-            receiver: abi.encode(address(s_sender)),
-            data: "",
-            tokenAmounts: destTokenAmounts,
-            feeToken: s_sourceFeeToken,
-            extraArgs: ""
-        });
+    Client.EVM2AnyMessage memory message = Client.EVM2AnyMessage({
+      receiver: abi.encode(address(s_sender)),
+      data: "",
+      tokenAmounts: destTokenAmounts,
+      feeToken: s_sourceFeeToken,
+      extraArgs: ""
+    });
 
-        uint256 feeTokenAmount = s_sourceRouter.getFee(DEST_CHAIN_SELECTOR, message);
-        uint256 feeTokenBalanceBefore = IERC20(s_sourceFeeToken).balanceOf(OWNER);
+    uint256 feeTokenAmount = s_sourceRouter.getFee(DEST_CHAIN_SELECTOR, message);
+    uint256 feeTokenBalanceBefore = IERC20(s_sourceFeeToken).balanceOf(OWNER);
 
-        s_sender.ccipSend({
-            destChainSelector: DEST_CHAIN_SELECTOR,
-            tokenAmounts: destTokenAmounts,
-            data: "",
-            feeToken: address(s_sourceFeeToken)
-        });
+    s_sender.ccipSend({
+      destChainSelector: DEST_CHAIN_SELECTOR,
+      tokenAmounts: destTokenAmounts,
+      data: "",
+      feeToken: address(s_sourceFeeToken)
+    });
 
-        // Assert that tokens were transfered for bridging + fees
-        assertEq(IERC20(token).balanceOf(OWNER), feeTokenBalanceBefore - amount - feeTokenAmount);
-    }
+    // Assert that tokens were transfered for bridging + fees
+    assertEq(IERC20(token).balanceOf(OWNER), feeTokenBalanceBefore - amount - feeTokenAmount);
+  }
 
-    function test_ccipSend_with_NativeFeeToken_andDestTokens() public {
-        address token = address(s_sourceFeeToken);
-        uint256 amount = 111333333777;
-        Client.EVMTokenAmount[] memory destTokenAmounts = new Client.EVMTokenAmount[](1);
-        destTokenAmounts[0] = Client.EVMTokenAmount({token: token, amount: amount});
+  function test_ccipSend_with_NativeFeeToken_andDestTokens() public {
+    address token = address(s_sourceFeeToken);
+    uint256 amount = 111333333777;
+    Client.EVMTokenAmount[] memory destTokenAmounts = new Client.EVMTokenAmount[](1);
+    destTokenAmounts[0] = Client.EVMTokenAmount({token: token, amount: amount});
 
-        // Make sure we give the receiver contract enough tokens like CCIP would.
-        IERC20(token).approve(address(s_sender), type(uint).max);
+    // Make sure we give the receiver contract enough tokens like CCIP would.
+    IERC20(token).approve(address(s_sender), type(uint256).max);
 
-        Client.EVM2AnyMessage memory message = Client.EVM2AnyMessage({
-            receiver: abi.encode(address(s_sender)),
-            data: "",
-            tokenAmounts: destTokenAmounts,
-            feeToken: address(0),
-            extraArgs: ""
-        });
+    Client.EVM2AnyMessage memory message = Client.EVM2AnyMessage({
+      receiver: abi.encode(address(s_sender)),
+      data: "",
+      tokenAmounts: destTokenAmounts,
+      feeToken: address(0),
+      extraArgs: ""
+    });
 
-        uint256 feeTokenAmount = s_sourceRouter.getFee(DEST_CHAIN_SELECTOR, message);
-        uint256 tokenBalanceBefore = IERC20(token).balanceOf(OWNER);
-        uint256 nativeFeeTokenBalanceBefore = OWNER.balance;
+    uint256 feeTokenAmount = s_sourceRouter.getFee(DEST_CHAIN_SELECTOR, message);
+    uint256 tokenBalanceBefore = IERC20(token).balanceOf(OWNER);
+    uint256 nativeFeeTokenBalanceBefore = OWNER.balance;
 
-        s_sender.ccipSend{value: feeTokenAmount}({
-            destChainSelector: DEST_CHAIN_SELECTOR,
-            tokenAmounts: destTokenAmounts,
-            data: "",
-            feeToken: address(0)
-        });
+    s_sender.ccipSend{value: feeTokenAmount}({
+      destChainSelector: DEST_CHAIN_SELECTOR,
+      tokenAmounts: destTokenAmounts,
+      data: "",
+      feeToken: address(0)
+    });
 
-        // Assert that native fees are paid successfully and tokens are transferred
-        assertEq(IERC20(token).balanceOf(OWNER), tokenBalanceBefore - amount);
-        assertEq(OWNER.balance, nativeFeeTokenBalanceBefore - feeTokenAmount);
-    
-    }
+    // Assert that native fees are paid successfully and tokens are transferred
+    assertEq(IERC20(token).balanceOf(OWNER), tokenBalanceBefore - amount);
+    assertEq(OWNER.balance, nativeFeeTokenBalanceBefore - feeTokenAmount);
+  }
 
-    function test_ccipSend_withNativeFeeToken_butInsufficientMsgValue_REVERT() public {
-        address token = address(s_sourceFeeToken);
-        uint256 amount = 111333333777;
-        Client.EVMTokenAmount[] memory destTokenAmounts = new Client.EVMTokenAmount[](0);
+  function test_ccipSend_withNativeFeeToken_butInsufficientMsgValue_REVERT() public {
+    address token = address(s_sourceFeeToken);
+    uint256 amount = 111333333777;
+    Client.EVMTokenAmount[] memory destTokenAmounts = new Client.EVMTokenAmount[](0);
 
-        Client.EVM2AnyMessage memory message = Client.EVM2AnyMessage({
-            receiver: abi.encode(address(s_sender)),
-            data: "FAKE_DATA",
-            tokenAmounts: destTokenAmounts,
-            feeToken: address(0),
-            extraArgs: ""
-        });
+    Client.EVM2AnyMessage memory message = Client.EVM2AnyMessage({
+      receiver: abi.encode(address(s_sender)),
+      data: "FAKE_DATA",
+      tokenAmounts: destTokenAmounts,
+      feeToken: address(0),
+      extraArgs: ""
+    });
 
-        uint256 feeTokenAmount = s_sourceRouter.getFee(DEST_CHAIN_SELECTOR, message);
+    uint256 feeTokenAmount = s_sourceRouter.getFee(DEST_CHAIN_SELECTOR, message);
 
-        vm.expectRevert(IRouterClient.InsufficientFeeTokenAmount.selector);
-        s_sender.ccipSend{value: feeTokenAmount / 2}({
-            destChainSelector: DEST_CHAIN_SELECTOR,
-            tokenAmounts: destTokenAmounts,
-            data: "",
-            feeToken: address(0)
-        });
-    }
+    vm.expectRevert(IRouterClient.InsufficientFeeTokenAmount.selector);
+    s_sender.ccipSend{value: feeTokenAmount / 2}({
+      destChainSelector: DEST_CHAIN_SELECTOR,
+      tokenAmounts: destTokenAmounts,
+      data: "",
+      feeToken: address(0)
+    });
+  }
 }
